Fix demo client CI checks
